clean up in lib/glcomp

See merge request graphviz/graphviz!1967

+++ lib/glcomp/glutils.c
@@ -30,9 +30,9 @@ int GetFixedOGLPos(int x, int y, float kts, GLfloat * X, GLfloat * Y,
     /* unused */
     (void)kts;
 
-    glColor4f((GLfloat) 0, (GLfloat) 0, (GLfloat) 0, (GLfloat) 0.001);
+    glColor4f(0.0f, 0.0f, 0.0f, 0.001f);
     glBegin(GL_POINTS);
-    glVertex3f((GLfloat) - 100.00, (GLfloat) - 100.00, (GLfloat) 1.00);
+    glVertex3f(-100.0f, -100.0f, 1.0f);
     glEnd();
 
     glGetDoublev(GL_MODELVIEW_MATRIX, modelview);
@@ -69,10 +69,10 @@ int GetOGLPosRef(int x, int y, float *X, float *Y, float *Z)
     glGetIntegerv(GL_VIEWPORT, viewport);
 
     //draw a point  to a not important location to get window coordinates
-    glColor4f((GLfloat) 0, (GLfloat) 0, (GLfloat) 0, (GLfloat) 0.001);
+    glColor4f(0.0f, 0.0f, 0.0f, 0.001f);
 
     glBegin(GL_POINTS);
-    glVertex3f(-100.00, -100.00, 0.00);
+    glVertex3f(-100.0f, -100.0f, 0.0f);
     glEnd();
     gluProject(-100.0, -100.0, 0.00, modelview, projection, viewport,
 	       &wwinX, &wwinY, &wwinZ);
@@ -107,10 +107,10 @@ float GetOGLDistance(int l)
     glGetIntegerv(GL_VIEWPORT, viewport);
 
     //draw a point  to a not important location to get window coordinates
-    glColor4f((GLfloat) 0, (GLfloat) 0, (GLfloat) 0, (GLfloat) 0.001);
+    glColor4f(0.0f, 0.0f, 0.0f, 0.001f);
 
     glBegin(GL_POINTS);
-    glVertex3f(10.00, 10.00, 1.00);
+    glVertex3f(10.0f, 10.0f, 1.0f);
     glEnd();
     gluProject(10.0, 10.0, 1.00, modelview, projection, viewport, &wwinX,
 	       &wwinY, &wwinZ);
@@ -192,9 +192,9 @@ int GetFixedOGLPoslocal(int x, int y, GLfloat * X, GLfloat * Y,
     glGetDoublev(GL_PROJECTION_MATRIX, projection);
     glGetIntegerv(GL_VIEWPORT, viewport);
 
-    glColor4f((GLfloat) 0, (GLfloat) 0, (GLfloat) 0, (GLfloat) 0.001);
+    glColor4f(0.0f, 0.0f, 0.0f, 0.001f);
     glBegin(GL_POINTS);
-    glVertex3f(10.00, 10.00, 0.00);
+    glVertex3f(10.0f, 10.0f, 0.0f);
     glEnd();
 
     gluProject(10.0, 10.0, 1.00, modelview, projection, viewport, &wwinX,
@@ -287,7 +287,7 @@ void glCompSelectionBox(glCompSet * s)
     w = s->mouse.dragX;
     h = s->mouse.dragY;
     printf("%f %f  %f  %f \n", x, y, w, h);
-    glColor4f(1, 1, 1, 1);
+    glColor4f(1.0f, 1.0f, 1.0f, 1.0f);
 
     glBegin(GL_POLYGON);
     glVertex2f(0, 0);
@@ -316,15 +316,15 @@ void glCompCalcWidget(glCompCommon * parent, glCompCommon * child,
 	return;
     }
     if (parent->borderType == glBorderNone)
-	borderWidth = 0;
+	borderWidth = 0.0f;
     else
 	borderWidth = parent->borderWidth;
     if (child->align != glAlignNone)	//if alignment, make sure width and height is no greater than parent
     {
 	if (child->width > parent->width)
-	    ref->width = parent->width - (float) 2.0 *borderWidth;
+	    ref->width = parent->width - 2.0f *borderWidth;
 	if (child->height > parent->height)
-	    ref->height = parent->height - (float) 2.0 *borderWidth;;
+	    ref->height = parent->height - 2.0f *borderWidth;;
 
     }
 
@@ -336,13 +336,13 @@ void glCompCalcWidget(glCompCommon * parent, glCompCommon * child,
     case glAlignLeft:
 	ref->pos.x = parent->refPos.x + borderWidth;
 	ref->pos.y = parent->refPos.y + borderWidth;
-	ref->height = parent->height - 2 * borderWidth;
+	ref->height = parent->height - 2.0f * borderWidth;
 	break;
     case glAlignRight:
 	ref->pos.x =
 	    parent->refPos.x + parent->width - child->width - borderWidth;
 	ref->pos.y = parent->refPos.y + borderWidth;
-	ref->height = parent->height - 2 * borderWidth;
+	ref->height = parent->height - 2.0f * borderWidth;
 	break;
 
     case glAlignTop:
@@ -350,19 +350,19 @@ void glCompCalcWidget(glCompCommon * parent, glCompCommon * child,
 	    parent->refPos.y + parent->height - child->height -
 	    borderWidth;
 	ref->pos.x = parent->refPos.x;
-	ref->width = parent->width - 2 * borderWidth;
+	ref->width = parent->width - 2.0f * borderWidth;
 	break;
 
     case glAlignBottom:
 	ref->pos.y = parent->refPos.y + borderWidth;
 	ref->pos.x = parent->refPos.x + borderWidth;
-	ref->width = parent->width - 2 * borderWidth;
+	ref->width = parent->width - 2.0f * borderWidth;
 	break;
     case glAlignParent:
 	ref->pos.y = parent->refPos.y + borderWidth;
 	ref->pos.x = parent->refPos.x + borderWidth;;
-	ref->width = parent->width - 2 * borderWidth;;
-	ref->height = parent->height - 2 * borderWidth;
+	ref->width = parent->width - 2.0f * borderWidth;;
+	ref->height = parent->height - 2.0f * borderWidth;
 	break;
     case glAlignCenter:
     case glAlignNone:
